+++ Chapter_2_Designing_Databricks_Day_One/Project: Favorita Store Sales - Time Series Forecasting/Downloading Sales Forecast Data.py
@@ -12,7 +12,7 @@
 
 # COMMAND ----------
 
-# MAGIC %run ../global-setup $project_name=favorita_forecasting $catalog=lakehouse_in_action
+# MAGIC %run ../global-setup $project_name=favorita_forecasting
 
 # COMMAND ----------
