+++ Chapter_4_Exploring_and_Cleaning_Toward_a_Silver_Layer/Project: Favorita Store Sales - Time Series Forecasting/Exploring Favorita Sales Data.py
@@ -12,7 +12,7 @@
 
 # COMMAND ----------
 
-# MAGIC %run ../../global-setup $project_name=favorita_forecasting $catalog=lakehouse_in_action
+# MAGIC %run ../../global-setup $project_name=favorita_forecasting
 
 # COMMAND ----------
