+++ Chapter_4_Exploring_and_Cleaning_Toward_a_Silver_Layer/Project: Favorita Store Sales - Time Series Forecasting/Oil data imputing.py
@@ -11,7 +11,7 @@
 
 # COMMAND ----------
 
-# MAGIC %run ../../global-setup $project_name=favorita_forecasting $catalog=lakehouse_in_action
+# MAGIC %run ../../global-setup $project_name=favorita_forecasting
 
 # COMMAND ----------
